changed tracing to include to split out local logger from tracer
