+++ Providers/nxOMSAutomationWorker/automationworker/worker/__init__.py
@@ -19,4 +19,5 @@ from worker.urllib2httpclient import *
 from worker.util import *
 from worker.workerexception import *
 from worker.diydirs import *
+from worker.locallogger import *
